bot/requirements.txt
python-telegram-bot==20.7
loguru
pytest
black
flake8
isort 